requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9